Starting Over, sort of
Started using CML to store molecules and use as the input for data
(instead of arbitrary strings and building my own parser).  Have also
switched over to using dictionaries to store the molecules/atoms/bonds
instead of nested matrices.  Have not yet decided on if
atoms/bonds/molecules will be classes or not