Finished basic unit testing
Finished implementing the unit testing of CMLParser and CMLBuilder.
Made some minor adjustments to those same classes so they worked
logically and parsed the data that needed to be parsed.

Need to begin work on explicit specs for my program - hopefully will
start that tomorrow.  Also going to start on constructing the CML
file(s) to represent the elements