Working on generating/locating significant paths
Its still a little tricky.  I need to adjust how I build an iterable
compound - probably going to need a generator since there will be a lot
of computations.  Also probably memoization.  This will also need
support for rings and for places where I care that it's CH2 and not CH1
(or something like that)