I hate documentation
I can safely remove a lot of the error checking on the lowest level
stuff.  None of these should ever be called by the user directly, and
the middle layer will make sure all of that is taken care of.
Done with documentation (except for the kivy gui, but that needs a
complete rewrite anyway)